Revert "Added abstract modifier; Fixed problem in val/var"
This reverts commit 7a091e8e52820c255b356d8dc2ca7cec8eb09cf7.
